@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, {useState} from 'react';
 import {connect} from 'react-redux';
 
 import Aux from '../../hoc/_Aux';
@@ -6,37 +6,40 @@ import classes from './Layout.module.css';
 import Toolbar from '../Navigation/Toolbar/Toolbar';
 import SideDrawer from '../Navigation/SideDrawer/SideDrawer';
 
-class Layout extends Component {
-    state = {
-        showSideDrawer: false
+const layout = props => {
+    const [sideDrawerVisible, setSideDrawerVisible] = useState(false);
+    // state = {
+    //     showSideDrawer: false
+    // };
+
+    const sideDrawerClosedHandler = () => {
+        setSideDrawerVisible(false);
+        // this.setState({showSideDrawer: false});
     };
 
-    sideDrawerClosedHandler = () => {
-        this.setState({showSideDrawer: false});
+    const sideDrawerToggleHandler = () => {
+      setSideDrawerVisible(!sideDrawerVisible);
+      //   this.setState((prevState) => {
+      //     return {showSideDrawer: !prevState.showSideDrawer};
+      // });
     };
 
-    sideDrawerToggleHandler = () => {
-        this.setState((prevState) => {
-          return {showSideDrawer: !prevState.showSideDrawer};
-      });
-    };
-
-    render() {
+    // render() {
         return (
           <Aux>
             <Toolbar
-                isAuth={this.props.isAuthenticated}
-                drawerToggleClicked={this.sideDrawerToggleHandler} />
+                isAuth={props.isAuthenticated}
+                drawerToggleClicked={sideDrawerToggleHandler} />
             <SideDrawer
-                isAuth={this.props.isAuthenticated}
-                open={this.state.showSideDrawer}
-                closed={this.sideDrawerClosedHandler}/>
+                isAuth={props.isAuthenticated}
+                open={sideDrawerVisible}
+                closed={sideDrawerClosedHandler}/>
             <main className={classes.Content}>
-                {this.props.children}
+                {props.children}
             </main>
           </Aux>
         );
-    };
+    // };
 };
 
 const mapStateToProps = state => {
@@ -45,4 +48,4 @@ const mapStateToProps = state => {
     };
 };
 
-export default connect(mapStateToProps)(Layout);
+export default connect(mapStateToProps)(layout);
