@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, {useState, useEffect} from 'react';
 import {connect} from 'react-redux';
 import {Redirect} from 'react-router-dom';
 
@@ -9,9 +9,8 @@ import classes from './Auth.module.css';
 import * as actions from '../../store/actions/index';
 import {updatedObject, checkValidity} from '../../shared/utility';
 
-class Auth extends Component {
-    state = {
-        controls: {
+const auth = props => {
+    const [authForm, setAuthForm] = useState({
             email: {
                 elementType: 'input',
                 elementConfig: {
@@ -40,46 +39,43 @@ class Auth extends Component {
                   valid: false,
                   touched: false
             },
-        },
-        isSignup: true
-    };
+        })
+        const [isSignup, setIsSignup] = useState(true);
 
-    componentDidMount() {
-        if (!this.props.buildingBurger && this.props.authRedirectPath !== '/') {
-            this.props.onSetAuthRedirectPath();
+    useEffect(() => {
+        if (!props.buildingBurger && props.authRedirectPath !== '/') {
+            props.onSetAuthRedirectPath();
         }
-    };
+    }, []);
+
     // added utility
     // using updatedObject() in shared folder for deep cloning
     // using checkValidity() in shared folder for validation
-    inputChangedHandler = (event, controlName) => {
-        const updatedControls = updatedObject(this.state.controls, {
-            [controlName]: updatedObject(this.state.controls[controlName], {
+    const inputChangedHandler = (event, controlName) => {
+        const updatedControls = updatedObject(authForm, {
+            [controlName]: updatedObject(authForm[controlName], {
                 value: event.target.value,
-                valid: checkValidity(event.target.value, this.state.controls[controlName]),
+                valid: checkValidity(event.target.value, authForm[controlName]),
                 touched: true
             })
         });
-        this.setState({controls: updatedControls})
+        setAuthForm(updatedControls);
     };
 
-    submitHandler = (event) => {
+    const submitHandler = (event) => {
         event.preventDefault();
-        this.props.onAuth(this.state.controls.email.value, this.state.controls.password.value, this.state.isSignup);
+        props.onAuth(authForm.email.value, authForm.password.value, isSignup);
     };
 
-    switchAuthModeHandler = () => {
-        this.setState(prevState => {
-            return {isSignup: !prevState.isSignup};
-        });
+    const switchAuthModeHandler = () => {
+        setIsSignup(!isSignup);
     };
 
-    render() {
         const formElementsArray = [];
-        for (let key in this.state.controls) {
+        for (let key in authForm) {
             formElementsArray.push({
                 id: key,
-                config: this.state.controls[key]
+                config: authForm[key]
             });
         };
 
@@ -92,40 +88,39 @@ class Auth extends Component {
                 invalid={!formElement.config.valid}
                 shouldValidate={formElement.config.validation}
                 touched={formElement.config.touched}
-                changed={(event) => this.inputChangedHandler(event, formElement.id)}/>
+                changed={(event) => inputChangedHandler(event, formElement.id)}/>
 
         ));
 
-        if (this.props.loading) {
+        if (props.loading) {
             form = <Spinner />
         };
 
         let errorMsg = null;
-        if (this.props.error) {
+        if (props.error) {
             errorMsg = (
-                <p>{this.props.error.message}</p>
+                <p>{props.error.message}</p>
             );
         };
 
         let authRedirect = null;
-        if (this.props.isAuthenticated) {
-            authRedirect = <Redirect to={this.props.authRedirectPath}/>
+        if (props.isAuthenticated) {
+            authRedirect = <Redirect to={props.authRedirectPath}/>
         };
 
         return (
             <div className={classes.Auth}>
                 {authRedirect}
                 {errorMsg}
-                <form onSubmit={this.submitHandler}>
+                <form onSubmit={submitHandler}>
                   {form}
                   <Button btnType="Success">Submit</Button>
                 </form>
                 <Button
-                    clicked={this.switchAuthModeHandler}
-                    btnType="Danger">Switch To {this.state.isSignup ? 'Sign In' : 'Sign Up'}</Button>
+                    clicked={switchAuthModeHandler}
+                    btnType="Danger">Switch To {isSignup ? 'Sign In' : 'Sign Up'}</Button>
             </div>
         );
-    };
 };
 
 const mapStateToProps = state => {
@@ -145,4 +140,4 @@ const mapDispatchToProps = dispatch => {
     };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(Auth);
+export default connect(mapStateToProps, mapDispatchToProps)(auth);
