@@ -1,17 +1,15 @@
-import React, {Component} from 'react';
+import React, {useEffect} from 'react';
 import {Redirect} from 'react-router-dom';
 import {connect} from 'react-redux';
 
 import * as actions from '../../../store/actions/index';
 
-class Logout extends Component {
-    componentDidMount() {
-        this.props.onLogout();
-    };
+const logout = props => {
+    useEffect(() => {
+        props.onLogout();
+    }, []);
 
-    render() {
-        return <Redirect to="/"/>;
-    };
+    return <Redirect to="/"/>;
 };
 
 const mapDispatchToProps = dispatch => {
@@ -20,4 +18,4 @@ const mapDispatchToProps = dispatch => {
     };
 };
 
-export default connect(null, mapDispatchToProps)(Logout);
+export default connect(null, mapDispatchToProps)(logout);
