@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, {useState} from 'react';
 import {connect} from 'react-redux';
 
 import Button from '../../../components/UI/Button/Button';
@@ -10,9 +10,8 @@ import withErrorHandler from '../../../hoc/withErrorHandler/withErrorHandler';
 import * as actions from '../../../store/actions/index';
 import {updatedObject, checkValidity} from '../../../shared/utility';
 
-class ContactData extends Component {
-    state = {
-        orderForm: {
+const contactData = props => {
+    const [orderForm, setOrderForm] = useState({
               name: {
                   elementType: 'input',
                   elementConfig: {
@@ -94,31 +93,29 @@ class ContactData extends Component {
                   validation: {},
                   valid: true
               }
-        },
-        formIsValid: false,
-        loading: false
-    };
+        })
+        const [formIsValid, setFormIsValid] = useState(false);
 
     // moved checkValidity() to shared folder. Added to Auth.js in Auth container
 
     //preventDefault using the event prop will stop sending a request and reloading the page
-    orderHandler = (event) => {
+    const orderHandler = (event) => {
         event.preventDefault();
         // this.setState({loading: true})
         const formData = {};
         // email, name, country... etc.
-        for (let formElementIdentifier in this.state.orderForm) {
-            formData[formElementIdentifier] = this.state.orderForm[formElementIdentifier].value;
+        for (let formElementIdentifier in orderForm) {
+            formData[formElementIdentifier] = orderForm[formElementIdentifier].value;
         };
         const order = {
             // add mapStateToProps
-            ingredients: this.props.ings,
-            price: this.props.price,
+            ingredients: props.ings,
+            price: props.price,
             orderData: formData,
-            userId: this.props.userId
+            userId: props.userId
         }
 
-        this.props.onOrderBurger(order, this.props.token);
+        props.onOrderBurger(order, props.token);
         // for firebase to work use node name.json
         // using it in actions now
         // axios.post('/orders.json', order)
@@ -135,33 +132,32 @@ class ContactData extends Component {
     // added utility
     // using updatedObject() in shared folder for deep cloning
     // using checkValidity() in shared folder for validation
-    inputChangedHandler = (event, id) => {
-        const updatedFormElement = updatedObject(this.state.orderForm[id], {
+    const inputChangedHandler = (event, id) => {
+        const updatedFormElement = updatedObject(orderForm[id], {
             value: event.target.value,
-            valid: checkValidity(event.target.value, this.state.orderForm[id].validation),
+            valid: checkValidity(event.target.value, orderForm[id].validation),
             touched: true
         });
-        const updatedOrderForm = updatedObject(this.state.orderForm, {
+        const updatedOrderForm = updatedObject(orderForm, {
             [id]: updatedFormElement
         });
         let formIsValid = true;
         for (let id in updatedOrderForm) {
             formIsValid = updatedOrderForm[id].valid && formIsValid;
         }
-
-        this.setState({orderForm: updatedOrderForm, formIsValid: formIsValid});
+        setOrderForm(updatedOrderForm);
+        setFormIsValid(formIsValid);
     };
 
-    render() {
         const formElementsArray = [];
-        for (let key in this.state.orderForm) {
+        for (let key in orderForm) {
             formElementsArray.push({
                 id: key,
-                config: this.state.orderForm[key]
+                config: orderForm[key]
             });
         };
         let form = (
-          <form onSubmit={this.orderHandler}>
+          <form onSubmit={orderHandler}>
               {formElementsArray.map(formElement => (
                   <Input
                       key={formElement.id}
@@ -171,13 +167,13 @@ class ContactData extends Component {
                       invalid={!formElement.config.valid}
                       shouldValidate={formElement.config.validation}
                       touched={formElement.config.touched}
-                      changed={(event) => this.inputChangedHandler(event, formElement.id)}/>
+                      changed={(event) => inputChangedHandler(event, formElement.id)}/>
               ))}
-              <Button btnType="Success" disabled={!this.state.formIsValid}>ORDER</Button>
+              <Button btnType="Success" disabled={!formIsValid}>ORDER</Button>
           </form>
         );
         // now using mapStateToProps instead
-        if (this.props.loading) {
+        if (props.loading) {
           form = <Spinner />;
         }
         return (
@@ -186,7 +182,6 @@ class ContactData extends Component {
                 {form}
             </div>
         );
-    };
 };
 
 const mapStateToProps = state => {
@@ -205,4 +200,4 @@ const mapDispatchToProps = dispatch => {
     };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(withErrorHandler(ContactData, axios));
+export default connect(mapStateToProps, mapDispatchToProps)(withErrorHandler(contactData, axios));
