@@ -1,7 +1,7 @@
-import React, { Component } from 'react';
+import React, {useEffect, Suspense } from 'react';
 import { Route, Switch, withRouter, Redirect } from 'react-router-dom';
 import {connect} from 'react-redux';
-import asyncComponent from './hoc/asyncComponent/asyncComponent';
+// import asyncComponent from './hoc/asyncComponent/asyncComponent';
 
 import Layout from './components/Layout/Layout';
 import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder';
@@ -10,40 +10,41 @@ import * as actions from './store/actions/index';
 
 // added lazy loading
 // only loads when needed
-const asyncCheckout = asyncComponent(() => {
+// React.lazy() can be used instead so we can get rid of the hoc asyncComponent file
+// wrap the functions with Suspense from React
+
+const Checkout = React.lazy(() => {
     return import ('./containers/Checkout/Checkout');
 });
 
-const asyncOrders = asyncComponent(() => {
+const Orders = React.lazy(() => {
     return import ('./containers/Orders/Orders');
 });
 
-const asyncAuth = asyncComponent(() => {
+const Auth = React.lazy(() => {
     return import ('./containers/Auth/Auth');
 });
 
+const app = props => {
+  useEffect(() => {
+    props.onTryAutoSignup();
+  }, []);
 
-class App extends Component {
-  componentDidMount() {
-      this.props.onTryAutoSignup();
-  };
-
-  render() {
     let routes = (
       <Switch>
-          <Route path="/auth" component={asyncAuth} />
+          <Route path="/auth" render={props => <Auth {...props}/>} />
           <Route path="/" exact component={BurgerBuilder} />
           <Redirect to="/" />
       </Switch>
     );
 
-    if (this.props.isAuthenticated) {
+    if (props.isAuthenticated) {
         routes = (
           <Switch>
-              <Route path="/checkout" component={asyncCheckout} />
-              <Route path="/orders" component={asyncOrders} />
+              <Route path="/checkout" render={props => <Checkout {...props}/>} />
+              <Route path="/orders" render={props => <Orders {...props}/>} />
               <Route path="/logout" component={Logout} />
-              <Route path="/auth" component={asyncAuth} />
+              <Route path="/auth" render={props => <Auth {...props}/>} />
               <Route path="/" exact component={BurgerBuilder} />
               <Redirect to="/" />
           </Switch>
@@ -53,11 +54,10 @@ class App extends Component {
     return (
       <div>
         <Layout>
-            {routes}
+            <Suspense fallback={<p>Loading...</p>}>{routes}</Suspense>
         </Layout>
       </div>
     );
-  };
 };
 
 const mapStateToProps = state => {
@@ -72,4 +72,4 @@ const mapDispatchToProps = dispatch => {
     };
 };
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(App));
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(app));
