@@ -1,15 +1,14 @@
-import React, {Component} from 'react';
+import React from 'react';
 import Aux from '../../../hoc/_Aux'
 import Button from '../../UI/Button/Button';
 
-class OrderSummary extends Component {
-    render() {
-      const ingredientSummary = Object.keys(this.props.ingredients)
+const orderSummary = props => {
+      const ingredientSummary = Object.keys(props.ingredients)
           .map(igKey => {
               return (
                 <li key={igKey}>
                   <span style={{textTransform: 'capitalize'}}>{igKey}</span>:
-                  {this.props.ingredients[igKey]}
+                  {props.ingredients[igKey]}
                 </li>
               );
           });
@@ -21,17 +20,16 @@ class OrderSummary extends Component {
               <ul>
                   {ingredientSummary}
               </ul>
-              <p><strong>Total Price: ${this.props.price.toFixed(2)}</strong></p>
+              <p><strong>Total Price: ${props.price.toFixed(2)}</strong></p>
               <p>Continue to Checkout?</p>
               <Button
                 btnType="Danger"
-                clicked={this.props.purchaseCancelled}>CANCEL</Button>
+                clicked={props.purchaseCancelled}>CANCEL</Button>
               <Button
                 btnType="Success"
-                clicked={this.props.purchaseContinue}>CONTINUE</Button>
+                clicked={props.purchaseContinue}>CONTINUE</Button>
           </Aux>
         );
-    };
 };
 
-export default OrderSummary;
+export default orderSummary;
