@@ -1,11 +1,11 @@
-import React, {Component} from 'react';
+import React, {} from 'react';
 import {Route, Redirect} from 'react-router-dom';
 import {connect} from 'react-redux';
 
 import CheckoutSummary from '../../components/Order/CheckoutSummary/CheckoutSummary';
 import ContactData from './ContactData/ContactData';
 
-class Checkout extends Component {
+const checkout = props => {
 
     // state = {
     //     ingredients: null,
@@ -28,29 +28,28 @@ class Checkout extends Component {
     //     }
     //     this.setState({ingredients: ingredients, totalPrice: price});
     // };
-    
-    checkoutCancelledHandler = () => {
-        this.props.history.goBack();
+
+    const checkoutCancelledHandler = () => {
+        props.history.goBack();
     };
 
-    checkoutContinuedHandler = () => {
-        this.props.history.replace('/checkout/contact-data');
+    const checkoutContinuedHandler = () => {
+        props.history.replace('/checkout/contact-data');
     };
 
     // Redirect to the homepage
-    render() {
         let summary = <Redirect to="/"/>
-        if (this.props.ings) {
-            const purchasedRedirect = this.props.purchased ? <Redirect to="/" /> : null;
+        if (props.ings) {
+            const purchasedRedirect = props.purchased ? <Redirect to="/" /> : null;
             summary = (
                 <div>
                     {purchasedRedirect}
                     <CheckoutSummary
-                        ingredients={this.props.ings}
-                        checkoutCancelled={this.checkoutCancelledHandler}
-                        checkoutContinued={this.checkoutContinuedHandler}/>
+                        ingredients={props.ings}
+                        checkoutCancelled={checkoutCancelledHandler}
+                        checkoutContinued={checkoutContinuedHandler}/>
                     <Route
-                        path={this.props.match.path + '/contact-data' }
+                        path={props.match.path + '/contact-data' }
                         component={ContactData} />
                         {/*render={(props) => (
                             <ContactData
@@ -61,7 +60,6 @@ class Checkout extends Component {
             );
         }
         return summary;
-    };
 };
 
 const mapStateToProps = state => {
@@ -71,4 +69,4 @@ const mapStateToProps = state => {
     };
 };
 
-export default connect(mapStateToProps)(Checkout);
+export default connect(mapStateToProps)(checkout);
